+++ auth/api/experimental/api_test.go
@@ -24,6 +24,7 @@ import (
 	"fmt"
 	"github.com/nuts-foundation/nuts-node/mock"
 	"github.com/nuts-foundation/nuts-node/vdr"
+	"github.com/sirupsen/logrus"
 	"net/http"
 	"reflect"
 	"testing"
@@ -45,6 +46,7 @@ type TestContext struct {
 	authMock           pkg2.AuthenticationServices
 	notaryMock         *services.MockContractNotary
 	contractClientMock *services.MockContractClient
+	oauthClientMock    *services.MockOAuthClient
 	wrapper            Wrapper
 }
 
@@ -52,10 +54,11 @@ type mockAuthClient struct {
 	ctrl               *gomock.Controller
 	mockContractClient *services.MockContractClient
 	mockContractNotary *services.MockContractNotary
+	mockOAuthClient    *services.MockOAuthClient
 }
 
 func (m *mockAuthClient) OAuthClient() services.OAuthClient {
-	panic("implement me")
+	return m.mockOAuthClient
 }
 
 func (m *mockAuthClient) ContractClient() services.ContractClient {
@@ -66,18 +69,20 @@ func (m *mockAuthClient) ContractNotary() services.ContractNotary {
 	return m.mockContractNotary
 }
 
-func createContext(t *testing.T) TestContext {
+func createContext(t *testing.T) *TestContext {
 	t.Helper()
 	ctrl := gomock.NewController(t)
 	mockContractClient := services.NewMockContractClient(ctrl)
 	mockContractNotary := services.NewMockContractNotary(ctrl)
-	authMock := &mockAuthClient{ctrl: ctrl, mockContractClient: mockContractClient, mockContractNotary: mockContractNotary}
-	return TestContext{
+	mockOAuthClient := services.NewMockOAuthClient(ctrl)
+	authMock := &mockAuthClient{ctrl: ctrl, mockContractClient: mockContractClient, mockContractNotary: mockContractNotary, mockOAuthClient: mockOAuthClient}
+	return &TestContext{
 		ctrl:               ctrl,
 		echoMock:           mock.NewMockContext(ctrl),
 		authMock:           authMock,
 		notaryMock:         mockContractNotary,
 		contractClientMock: mockContractClient,
+		oauthClientMock:    mockOAuthClient,
 		wrapper:            Wrapper{Auth: authMock},
 	}
 }
@@ -172,7 +177,6 @@ func TestWrapper_GetSignSessionStatus(t *testing.T) {
 	})
 }
 
-
 func TestWrapper_GetContractByType(t *testing.T) {
 	t.Run("get known contact", func(t *testing.T) {
 		ctx := createContext(t)
@@ -238,7 +242,7 @@ func TestWrapper_DrawUpContract(t *testing.T) {
 			Version:     ContractVersion("v3"),
 			LegalEntity: LegalEntity(vdr.TestDIDA.String()),
 		}
-		bindPostBody(&ctx, params)
+		bindPostBody(ctx, params)
 
 		template := contract.StandardContractTemplates["EN"]["PractitionerLogin"]["v3"]
 		drawnUpContract := &contract.Contract{
@@ -269,7 +273,7 @@ func TestWrapper_DrawUpContract(t *testing.T) {
 			params := DrawUpContractRequest{
 				ValidFrom: &validFrom,
 			}
-			bindPostBody(&ctx, params)
+			bindPostBody(ctx, params)
 
 			err := ctx.wrapper.DrawUpContract(ctx.echoMock)
 
@@ -288,7 +292,7 @@ func TestWrapper_DrawUpContract(t *testing.T) {
 			params := DrawUpContractRequest{
 				ValidDuration: &duration,
 			}
-			bindPostBody(&ctx, params)
+			bindPostBody(ctx, params)
 
 			err := ctx.wrapper.DrawUpContract(ctx.echoMock)
 
@@ -307,7 +311,7 @@ func TestWrapper_DrawUpContract(t *testing.T) {
 				Type:     ContractType("UnknownContractName"),
 				Version:  ContractVersion("v3"),
 			}
-			bindPostBody(&ctx, params)
+			bindPostBody(ctx, params)
 
 			err := ctx.wrapper.DrawUpContract(ctx.echoMock)
 
@@ -327,7 +331,7 @@ func TestWrapper_DrawUpContract(t *testing.T) {
 				Version:     ContractVersion("v3"),
 				LegalEntity: LegalEntity("ZorgId:15"),
 			}
-			bindPostBody(&ctx, params)
+			bindPostBody(ctx, params)
 
 			err := ctx.wrapper.DrawUpContract(ctx.echoMock)
 
@@ -349,7 +353,7 @@ func TestWrapper_DrawUpContract(t *testing.T) {
 			Version:     ContractVersion("v3"),
 			LegalEntity: LegalEntity(vdr.TestDIDA.String()),
 		}
-		bindPostBody(&ctx, params)
+		bindPostBody(ctx, params)
 
 		ctx.notaryMock.EXPECT().DrawUpContract(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).Return(nil, errors.New("unknown error while drawing up the contract"))
 
@@ -360,6 +364,279 @@ func TestWrapper_DrawUpContract(t *testing.T) {
 		assert.Equal(t, http.StatusBadRequest, httpError.Code)
 		assert.Equal(t, "error while drawing up the contract: unknown error while drawing up the contract", httpError.Message)
 	})
+}
+
+func TestWrapper_CreateJwtBearerToken(t *testing.T) {
+	bindPostBody := func(ctx *TestContext, body CreateJwtBearerTokenRequest) {
+		jsonData, _ := json.Marshal(body)
+		ctx.echoMock.EXPECT().Bind(gomock.Any()).Do(func(f interface{}) {
+			_ = json.Unmarshal(jsonData, f)
+		})
+	}
+
+	expectStatusOK := func(ctx *TestContext, response JwtBearerTokenResponse) {
+		ctx.echoMock.EXPECT().JSON(http.StatusOK, response)
+	}
+
+	t.Run("make request", func(t *testing.T) {
+		ctx := createContext(t)
+		defer ctx.ctrl.Finish()
+		subj := "urn:oid:2.16.840.1.113883.2.4.6.3:9999990"
+		body := CreateJwtBearerTokenRequest{
+			Actor:     "urn:oid:2.16.840.1.113883.2.4.6.1:48000000",
+			Custodian: "urn:oid:2.16.840.1.113883.2.4.6.1:12481248",
+			Subject:   &subj,
+			Identity:  "irma-token",
+			Scope:     "nuts-sso",
+		}
+		bindPostBody(ctx, body)
+		response := JwtBearerTokenResponse{
+			BearerToken: "123.456.789",
+		}
+
+		expectedRequest := services.CreateJwtBearerTokenRequest{
+			Actor:         body.Actor,
+			Custodian:     body.Custodian,
+			IdentityToken: &body.Identity,
+			Subject:       body.Subject,
+		}
+
+		ctx.oauthClientMock.EXPECT().CreateJwtBearerToken(expectedRequest).Return(&services.JwtBearerTokenResult{BearerToken: response.BearerToken}, nil)
+		expectStatusOK(ctx, response)
+
+		if !assert.Nil(t, ctx.wrapper.CreateJwtBearerToken(ctx.echoMock)) {
+			t.FailNow()
+		}
+	})
+}
+
+func TestWrapper_CreateAccessToken(t *testing.T) {
+	const validJwt = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJ1cm46b2lkOjIuMTYuODQwLjEuMTEzODgzLjIuNC42LjE6NDgwMDAwMDAiLCJzdWIiOiJ1cm46b2lkOjIuMTYuODQwLjEuMTEzODgzLjIuNC42LjE6MTI0ODEyNDgiLCJzaWQiOiJ1cm46b2lkOjIuMTYuODQwLjEuMTEzODgzLjIuNC42LjM6OTk5OTk5MCIsImF1ZCI6Imh0dHBzOi8vdGFyZ2V0X3Rva2VuX2VuZHBvaW50IiwidXNpIjoiYmFzZTY0IGVuY29kZWQgc2lnbmF0dXJlIiwiZXhwIjoxNTc4MTEwNDgxLCJpYXQiOjE1Nzg5MTA0ODEsImp0aSI6IjEyMy00NTYtNzg5In0.76XtU81IyR3Ak_2fgrYsuLcvxndf0eedT1mFPa-rPXk"
+
+	bindPostBody := func(ctx *TestContext, body CreateAccessTokenRequest) {
+		ctx.echoMock.EXPECT().FormValue("assertion").Return(body.Assertion)
+		ctx.echoMock.EXPECT().FormValue("grant_type").Return(body.GrantType)
+	}
+
+	expectError := func(ctx *TestContext, err AccessTokenRequestFailedResponse) {
+		ctx.echoMock.EXPECT().JSON(http.StatusBadRequest, oauthErrorMatcher{x: err})
+	}
+
+	expectStatusOK := func(ctx *TestContext, response AccessTokenResponse) {
+		ctx.echoMock.EXPECT().JSON(http.StatusOK, gomock.Eq(response))
+	}
+
+	t.Run("unknown grant_type", func(t *testing.T) {
+		ctx := createContext(t)
+		defer ctx.ctrl.Finish()
+
+		params := CreateAccessTokenRequest{GrantType: "unknown type"}
+		bindPostBody(ctx, params)
+
+		errorDescription := "grant_type must be: 'urn:ietf:params:oauth:grant-type:jwt-bearer'"
+		errorResponse := AccessTokenRequestFailedResponse{ErrorDescription: errorDescription, Error: errOauthUnsupportedGrant}
+		expectError(ctx, errorResponse)
+
+		err := ctx.wrapper.CreateAccessToken(ctx.echoMock, CreateAccessTokenParams{})
+
+		assert.Nil(t, err)
+	})
+
+	t.Run("invalid assertion", func(t *testing.T) {
+		ctx := createContext(t)
+		defer ctx.ctrl.Finish()
+
+		params := CreateAccessTokenRequest{GrantType: "urn:ietf:params:oauth:grant-type:jwt-bearer", Assertion: "invalid jwt"}
+		bindPostBody(ctx, params)
+
+		errorDescription := "Assertion must be a valid encoded jwt"
+		errorResponse := AccessTokenRequestFailedResponse{ErrorDescription: errorDescription, Error: errOauthInvalidGrant}
+		expectError(ctx, errorResponse)
+
+		err := ctx.wrapper.CreateAccessToken(ctx.echoMock, CreateAccessTokenParams{})
+
+		assert.Nil(t, err)
+	})
+
+	t.Run("auth.CreateAccessToken returns error", func(t *testing.T) {
+		ctx := createContext(t)
+		defer ctx.ctrl.Finish()
+
+		params := CreateAccessTokenRequest{GrantType: "urn:ietf:params:oauth:grant-type:jwt-bearer", Assertion: validJwt}
+		bindPostBody(ctx, params)
+
+		errorDescription := "oh boy"
+		errorResponse := AccessTokenRequestFailedResponse{ErrorDescription: errorDescription, Error: errOauthInvalidRequest}
+		expectError(ctx, errorResponse)
+
+		ctx.oauthClientMock.EXPECT().CreateAccessToken(services.CreateAccessTokenRequest{RawJwtBearerToken: validJwt}).Return(nil, fmt.Errorf("oh boy"))
+		err := ctx.wrapper.CreateAccessToken(ctx.echoMock, CreateAccessTokenParams{XSslClientCert: "cert"})
+
+		assert.Nil(t, err)
+	})
+
+	t.Run("valid request", func(t *testing.T) {
+		ctx := createContext(t)
+		defer ctx.ctrl.Finish()
+
+		params := CreateAccessTokenRequest{GrantType: "urn:ietf:params:oauth:grant-type:jwt-bearer", Assertion: validJwt}
+		bindPostBody(ctx, params)
+
+		pkgResponse := &services.AccessTokenResult{AccessToken: "foo"}
+		ctx.oauthClientMock.EXPECT().CreateAccessToken(services.CreateAccessTokenRequest{RawJwtBearerToken: validJwt}).Return(pkgResponse, nil)
+
+		apiResponse := AccessTokenResponse{AccessToken: pkgResponse.AccessToken}
+		expectStatusOK(ctx, apiResponse)
+
+		err := ctx.wrapper.CreateAccessToken(ctx.echoMock, CreateAccessTokenParams{})
+
+		assert.Nil(t, err)
+
+	})
+}
+
+func TestWrapper_VerifyAccessToken(t *testing.T) {
+	t.Run("403 - missing header", func(t *testing.T) {
+		ctx := createContext(t)
+		defer ctx.ctrl.Finish()
+		params := VerifyAccessTokenParams{
+			Authorization: "",
+		}
+
+		ctx.echoMock.EXPECT().NoContent(http.StatusForbidden)
+
+		_ = ctx.wrapper.VerifyAccessToken(ctx.echoMock, params)
+	})
+
+	t.Run("403 - incorrect authorization header", func(t *testing.T) {
+		ctx := createContext(t)
+		defer ctx.ctrl.Finish()
+		params := VerifyAccessTokenParams{
+			Authorization: "34987569ytihua",
+		}
+
+		ctx.echoMock.EXPECT().NoContent(http.StatusForbidden)
+
+		_ = ctx.wrapper.VerifyAccessToken(ctx.echoMock, params)
+	})
+
+	t.Run("403 - incorrect token", func(t *testing.T) {
+		ctx := createContext(t)
+		defer ctx.ctrl.Finish()
+		params := VerifyAccessTokenParams{
+			Authorization: "Bearer token",
+		}
+
+		ctx.echoMock.EXPECT().NoContent(http.StatusForbidden)
+		ctx.oauthClientMock.EXPECT().IntrospectAccessToken("token").Return(nil, errors.New("unauthorized"))
+
+		_ = ctx.wrapper.VerifyAccessToken(ctx.echoMock, params)
+	})
+
+	t.Run("200 - correct token", func(t *testing.T) {
+		ctx := createContext(t)
+		defer ctx.ctrl.Finish()
+		params := VerifyAccessTokenParams{
+			Authorization: "Bearer token",
+		}
+
+		ctx.echoMock.EXPECT().NoContent(http.StatusOK)
+		ctx.oauthClientMock.EXPECT().IntrospectAccessToken("token").Return(&services.NutsAccessToken{}, nil)
+
+		_ = ctx.wrapper.VerifyAccessToken(ctx.echoMock, params)
+	})
+}
+
+func TestWrapper_IntrospectAccessToken(t *testing.T) {
+	bindPostBody := func(ctx *TestContext, body TokenIntrospectionRequest) {
+		ctx.echoMock.EXPECT().FormValue("token").Return(body.Token)
+	}
+
+	expectStatusOK := func(ctx *TestContext, response TokenIntrospectionResponse) {
+		data, _ := json.Marshal(response)
+		logrus.Infof("Expect: %s", string(data))
+		ctx.echoMock.EXPECT().JSON(http.StatusOK, gomock.Eq(response))
+	}
+
+	t.Run("empty token returns active false", func(t *testing.T) {
+		ctx := createContext(t)
+		defer ctx.ctrl.Finish()
+
+		request := TokenIntrospectionRequest{Token: ""}
+		bindPostBody(ctx, request)
+
+		response := TokenIntrospectionResponse{Active: false}
+		expectStatusOK(ctx, response)
+
+		_ = ctx.wrapper.IntrospectAccessToken(ctx.echoMock)
+	})
+
+	t.Run("introspect a token", func(t *testing.T) {
+		ctx := createContext(t)
+		defer ctx.ctrl.Finish()
+
+		request := TokenIntrospectionRequest{Token: "123"}
+		bindPostBody(ctx, request)
+
+		aud := "123"
+		aid := "urn:oid:2.16.840.1.113883.2.4.6.1:00000000"
+		exp := 1581412667
+		iat := 1581411767
+		iss := "urn:oid:2.16.840.1.113883.2.4.6.1:00000001"
+		sid := "urn:oid:2.16.840.1.113883.2.4.6.3:999999990"
+		scope := "nuts-sso"
+
+		ctx.oauthClientMock.EXPECT().IntrospectAccessToken(request.Token).Return(
+			&services.NutsAccessToken{
+				Audience:   aud,
+				Expiration: int64(exp),
+				IssuedAt:   int64(iat),
+				Issuer:     iss,
+				Subject:    aid,
+				SubjectID:  &sid,
+				Scope:      scope,
+			}, nil)
+
+		emptyStr := ""
+		response := TokenIntrospectionResponse{
+			Active: true,
+			Aud:    &aud,
+			Exp:    &exp,
+			Iat:    &iat,
+			Iss:    &iss,
+			Sid:    &sid,
+			Sub:    &aid,
+			//Uid:    &uid,
+			Scope:      &scope,
+			Email:      &emptyStr,
+			GivenName:  &emptyStr,
+			Prefix:     &emptyStr,
+			FamilyName: &emptyStr,
+			Name:       &emptyStr,
+		}
+		expectStatusOK(ctx, response)
+
+		if !assert.NoError(t, ctx.wrapper.IntrospectAccessToken(ctx.echoMock)) {
+			t.Fail()
+		}
+	})
+}
+
+type oauthErrorMatcher struct {
+	x AccessTokenRequestFailedResponse
+}
+
+func (e oauthErrorMatcher) Matches(x interface{}) bool {
+	if !reflect.TypeOf(x).AssignableTo(reflect.TypeOf(x)) {
+		return false
+	}
+
+	response := x.(AccessTokenRequestFailedResponse)
+	return e.x.Error == response.Error && e.x.ErrorDescription == response.ErrorDescription
+}
+
+func (e oauthErrorMatcher) String() string {
+	return fmt.Sprintf("is equal to {%v, %v}", e.x.Error, e.x.ErrorDescription)
 }
 
 type signSessionResponseMatcher struct {
@@ -405,7 +682,7 @@ func TestWrapper_CreateSignSession(t *testing.T) {
 			Means:   "dummy",
 			Payload: "this is the contract message to agree to",
 		}
-		bindPostBody(&ctx, postParams)
+		bindPostBody(ctx, postParams)
 
 		ctx.echoMock.EXPECT().JSON(http.StatusCreated, signSessionResponseMatcher{means: "dummy"})
 		err := ctx.wrapper.CreateSignSession(ctx.echoMock)
@@ -417,7 +694,7 @@ func TestWrapper_CreateSignSession(t *testing.T) {
 		defer ctx.ctrl.Finish()
 
 		postParams := CreateSignSessionRequest{}
-		bindPostBody(&ctx, postParams)
+		bindPostBody(ctx, postParams)
 
 		ctx.contractClientMock.EXPECT().CreateSigningSession(gomock.Any()).Return(nil, errors.New("some error"))
 
@@ -449,7 +726,7 @@ func TestWrapper_VerifySignature(t *testing.T) {
 				Type:    []string{"TestCredential"},
 			}}
 
-		bindPostBody(&ctx, postParams)
+		bindPostBody(ctx, postParams)
 
 		verificationResult := &contract.VPVerificationResult{
 			Validity:            contract.Valid,
@@ -483,7 +760,7 @@ func TestWrapper_VerifySignature(t *testing.T) {
 		postParams := SignatureVerificationRequest{
 			VerifiablePresentation: VerifiablePresentation{}}
 
-		bindPostBody(&ctx, postParams)
+		bindPostBody(ctx, postParams)
 
 		verificationResult := &contract.VPVerificationResult{
 			Validity: contract.Invalid,
@@ -513,7 +790,7 @@ func TestWrapper_VerifySignature(t *testing.T) {
 				Type:    []string{"TestCredential"},
 			}}
 
-		bindPostBody(&ctx, postParams)
+		bindPostBody(ctx, postParams)
 
 		verificationResult := &contract.VPVerificationResult{
 			Validity: contract.Valid,
@@ -552,7 +829,7 @@ func TestWrapper_VerifySignature(t *testing.T) {
 			VerifiablePresentation: VerifiablePresentation{},
 		}
 
-		bindPostBody(&ctx, postParams)
+		bindPostBody(ctx, postParams)
 
 		err := ctx.wrapper.VerifySignature(ctx.echoMock)
 		if !assert.Error(t, err) {
@@ -569,7 +846,7 @@ func TestWrapper_VerifySignature(t *testing.T) {
 			VerifiablePresentation: VerifiablePresentation{},
 		}
 
-		bindPostBody(&ctx, postParams)
+		bindPostBody(ctx, postParams)
 
 		ctx.contractClientMock.EXPECT().VerifyVP(gomock.Any(), gomock.Any()).Return(nil, errors.New("verification error"))
