+++ auth/api/v0/api_test.go
@@ -19,39 +19,16 @@
 package v0
 
 import (
-	"encoding/json"
-	"errors"
-	"fmt"
-	"net/http"
-	"reflect"
 	"testing"
 
+	"github.com/golang/mock/gomock"
 	"github.com/nuts-foundation/nuts-node/auth"
 	"github.com/nuts-foundation/nuts-node/auth/services"
 	"github.com/nuts-foundation/nuts-node/mock"
-	"github.com/sirupsen/logrus"
-
-	"github.com/golang/mock/gomock"
-	"github.com/stretchr/testify/assert"
 )
 
 
-type OAuthErrorMatcher struct {
-	x AccessTokenRequestFailedResponse
-}
 
-func (e OAuthErrorMatcher) Matches(x interface{}) bool {
-	if !reflect.TypeOf(x).AssignableTo(reflect.TypeOf(x)) {
-		return false
-	}
-
-	response := x.(AccessTokenRequestFailedResponse)
-	return e.x.Error == response.Error && e.x.ErrorDescription == response.ErrorDescription
-}
-
-func (e OAuthErrorMatcher) String() string {
-	return fmt.Sprintf("is equal to {%v, %v}", e.x.Error, e.x.ErrorDescription)
-}
 
 type TestContext struct {
 	ctrl         *gomock.Controller
@@ -83,261 +60,4 @@ var createContext = func(t *testing.T) *TestContext {
 		notaryMock:   notaryMock,
 		wrapper:      Wrapper{Auth: authMock},
 	}
-}
-
-func TestWrapper_NutsAuthCreateAccessToken(t *testing.T) {
-	const validJwt = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJ1cm46b2lkOjIuMTYuODQwLjEuMTEzODgzLjIuNC42LjE6NDgwMDAwMDAiLCJzdWIiOiJ1cm46b2lkOjIuMTYuODQwLjEuMTEzODgzLjIuNC42LjE6MTI0ODEyNDgiLCJzaWQiOiJ1cm46b2lkOjIuMTYuODQwLjEuMTEzODgzLjIuNC42LjM6OTk5OTk5MCIsImF1ZCI6Imh0dHBzOi8vdGFyZ2V0X3Rva2VuX2VuZHBvaW50IiwidXNpIjoiYmFzZTY0IGVuY29kZWQgc2lnbmF0dXJlIiwiZXhwIjoxNTc4MTEwNDgxLCJpYXQiOjE1Nzg5MTA0ODEsImp0aSI6IjEyMy00NTYtNzg5In0.76XtU81IyR3Ak_2fgrYsuLcvxndf0eedT1mFPa-rPXk"
-
-	bindPostBody := func(ctx *TestContext, body CreateAccessTokenRequest) {
-		ctx.echoMock.EXPECT().FormValue("assertion").Return(body.Assertion)
-		ctx.echoMock.EXPECT().FormValue("grant_type").Return(body.GrantType)
-	}
-
-	expectError := func(ctx *TestContext, err AccessTokenRequestFailedResponse) {
-		ctx.echoMock.EXPECT().JSON(http.StatusBadRequest, OAuthErrorMatcher{x: err})
-	}
-
-	expectStatusOK := func(ctx *TestContext, response AccessTokenResponse) {
-		ctx.echoMock.EXPECT().JSON(http.StatusOK, gomock.Eq(response))
-	}
-
-	t.Run("unknown grant_type", func(t *testing.T) {
-		ctx := createContext(t)
-		defer ctx.ctrl.Finish()
-
-		params := CreateAccessTokenRequest{GrantType: "unknown type"}
-		bindPostBody(ctx, params)
-
-		errorDescription := "grant_type must be: 'urn:ietf:params:oauth:grant-type:jwt-bearer'"
-		errorResponse := AccessTokenRequestFailedResponse{ErrorDescription: errorDescription, Error: errOauthUnsupportedGrant}
-		expectError(ctx, errorResponse)
-
-		err := ctx.wrapper.CreateAccessToken(ctx.echoMock, CreateAccessTokenParams{})
-
-		assert.Nil(t, err)
-	})
-
-	t.Run("invalid assertion", func(t *testing.T) {
-		ctx := createContext(t)
-		defer ctx.ctrl.Finish()
-
-		params := CreateAccessTokenRequest{GrantType: "urn:ietf:params:oauth:grant-type:jwt-bearer", Assertion: "invalid jwt"}
-		bindPostBody(ctx, params)
-
-		errorDescription := "Assertion must be a valid encoded jwt"
-		errorResponse := AccessTokenRequestFailedResponse{ErrorDescription: errorDescription, Error: errOauthInvalidGrant}
-		expectError(ctx, errorResponse)
-
-		err := ctx.wrapper.CreateAccessToken(ctx.echoMock, CreateAccessTokenParams{})
-
-		assert.Nil(t, err)
-	})
-
-	t.Run("auth.CreateAccessToken returns error", func(t *testing.T) {
-		ctx := createContext(t)
-		defer ctx.ctrl.Finish()
-
-		params := CreateAccessTokenRequest{GrantType: "urn:ietf:params:oauth:grant-type:jwt-bearer", Assertion: validJwt}
-		bindPostBody(ctx, params)
-
-		errorDescription := "oh boy"
-		errorResponse := AccessTokenRequestFailedResponse{ErrorDescription: errorDescription, Error: errOauthInvalidRequest}
-		expectError(ctx, errorResponse)
-
-		ctx.oauthMock.EXPECT().CreateAccessToken(services.CreateAccessTokenRequest{RawJwtBearerToken: validJwt}).Return(nil, fmt.Errorf("oh boy"))
-		err := ctx.wrapper.CreateAccessToken(ctx.echoMock, CreateAccessTokenParams{XSslClientCert: "cert"})
-
-		assert.Nil(t, err)
-	})
-
-	t.Run("valid request", func(t *testing.T) {
-		ctx := createContext(t)
-		defer ctx.ctrl.Finish()
-
-		params := CreateAccessTokenRequest{GrantType: "urn:ietf:params:oauth:grant-type:jwt-bearer", Assertion: validJwt}
-		bindPostBody(ctx, params)
-
-		pkgResponse := &services.AccessTokenResult{AccessToken: "foo"}
-		ctx.oauthMock.EXPECT().CreateAccessToken(services.CreateAccessTokenRequest{RawJwtBearerToken: validJwt}).Return(pkgResponse, nil)
-
-		apiResponse := AccessTokenResponse{AccessToken: pkgResponse.AccessToken}
-		expectStatusOK(ctx, apiResponse)
-
-		err := ctx.wrapper.CreateAccessToken(ctx.echoMock, CreateAccessTokenParams{})
-
-		assert.Nil(t, err)
-
-	})
-
-}
-
-func TestWrapper_NutsAuthCreateJwtBearerToken(t *testing.T) {
-	bindPostBody := func(ctx *TestContext, body CreateJwtBearerTokenRequest) {
-		jsonData, _ := json.Marshal(body)
-		ctx.echoMock.EXPECT().Bind(gomock.Any()).Do(func(f interface{}) {
-			_ = json.Unmarshal(jsonData, f)
-		})
-	}
-
-	expectStatusOK := func(ctx *TestContext, response JwtBearerTokenResponse) {
-		ctx.echoMock.EXPECT().JSON(http.StatusOK, response)
-	}
-
-	t.Run("make request", func(t *testing.T) {
-		ctx := createContext(t)
-		defer ctx.ctrl.Finish()
-		subj := "urn:oid:2.16.840.1.113883.2.4.6.3:9999990"
-		body := CreateJwtBearerTokenRequest{
-			Actor:     "urn:oid:2.16.840.1.113883.2.4.6.1:48000000",
-			Custodian: "urn:oid:2.16.840.1.113883.2.4.6.1:12481248",
-			Subject:   &subj,
-			Identity:  "irma-token",
-			Scope:     "nuts-sso",
-		}
-		bindPostBody(ctx, body)
-		response := JwtBearerTokenResponse{
-			BearerToken: "123.456.789",
-		}
-
-		expectedRequest := services.CreateJwtBearerTokenRequest{
-			Actor:         body.Actor,
-			Custodian:     body.Custodian,
-			IdentityToken: &body.Identity,
-			Subject:       body.Subject,
-		}
-
-		ctx.oauthMock.EXPECT().CreateJwtBearerToken(expectedRequest).Return(&services.JwtBearerTokenResult{BearerToken: response.BearerToken}, nil)
-		expectStatusOK(ctx, response)
-
-		if !assert.Nil(t, ctx.wrapper.CreateJwtBearerToken(ctx.echoMock)) {
-			t.FailNow()
-		}
-	})
-}
-
-func TestWrapper_NutsAuthIntrospectAccessToken(t *testing.T) {
-	bindPostBody := func(ctx *TestContext, body TokenIntrospectionRequest) {
-		ctx.echoMock.EXPECT().FormValue("token").Return(body.Token)
-	}
-
-	expectStatusOK := func(ctx *TestContext, response TokenIntrospectionResponse) {
-		data, _ := json.Marshal(response)
-		logrus.Infof("Expect: %s", string(data))
-		ctx.echoMock.EXPECT().JSON(http.StatusOK, gomock.Eq(response))
-	}
-
-	t.Run("empty token returns active false", func(t *testing.T) {
-		ctx := createContext(t)
-		defer ctx.ctrl.Finish()
-
-		request := TokenIntrospectionRequest{Token: ""}
-		bindPostBody(ctx, request)
-
-		response := TokenIntrospectionResponse{Active: false}
-		expectStatusOK(ctx, response)
-
-		_ = ctx.wrapper.IntrospectAccessToken(ctx.echoMock)
-	})
-
-	t.Run("introspect a token", func(t *testing.T) {
-		ctx := createContext(t)
-		defer ctx.ctrl.Finish()
-
-		request := TokenIntrospectionRequest{Token: "123"}
-		bindPostBody(ctx, request)
-
-		aud := "123"
-		aid := "urn:oid:2.16.840.1.113883.2.4.6.1:00000000"
-		exp := 1581412667
-		iat := 1581411767
-		iss := "urn:oid:2.16.840.1.113883.2.4.6.1:00000001"
-		sid := "urn:oid:2.16.840.1.113883.2.4.6.3:999999990"
-		scope := "nuts-sso"
-
-		ctx.oauthMock.EXPECT().IntrospectAccessToken(request.Token).Return(
-			&services.NutsAccessToken{
-				Audience:   aud,
-				Expiration: int64(exp),
-				IssuedAt:   int64(iat),
-				Issuer:     iss,
-				Subject:    aid,
-				SubjectID:  &sid,
-				Scope:      scope,
-			}, nil)
-
-		emptyStr := ""
-		response := TokenIntrospectionResponse{
-			Active: true,
-			Aud:    &aud,
-			Exp:    &exp,
-			Iat:    &iat,
-			Iss:    &iss,
-			Sid:    &sid,
-			Sub:    &aid,
-			//Uid:    &uid,
-			Scope:      &scope,
-			Email:      &emptyStr,
-			GivenName:  &emptyStr,
-			Prefix:     &emptyStr,
-			FamilyName: &emptyStr,
-			Name:       &emptyStr,
-		}
-		expectStatusOK(ctx, response)
-
-		if !assert.NoError(t, ctx.wrapper.IntrospectAccessToken(ctx.echoMock)) {
-			t.Fail()
-		}
-	})
-}
-
-func TestWrapper_VerifyAccessToken(t *testing.T) {
-	t.Run("403 - missing header", func(t *testing.T) {
-		ctx := createContext(t)
-		defer ctx.ctrl.Finish()
-		params := VerifyAccessTokenParams{
-			Authorization: "",
-		}
-
-		ctx.echoMock.EXPECT().NoContent(http.StatusForbidden)
-
-		_ = ctx.wrapper.VerifyAccessToken(ctx.echoMock, params)
-	})
-
-	t.Run("403 - incorrect authorization header", func(t *testing.T) {
-		ctx := createContext(t)
-		defer ctx.ctrl.Finish()
-		params := VerifyAccessTokenParams{
-			Authorization: "34987569ytihua",
-		}
-
-		ctx.echoMock.EXPECT().NoContent(http.StatusForbidden)
-
-		_ = ctx.wrapper.VerifyAccessToken(ctx.echoMock, params)
-	})
-
-	t.Run("403 - incorrect token", func(t *testing.T) {
-		ctx := createContext(t)
-		defer ctx.ctrl.Finish()
-		params := VerifyAccessTokenParams{
-			Authorization: "Bearer token",
-		}
-
-		ctx.echoMock.EXPECT().NoContent(http.StatusForbidden)
-		ctx.oauthMock.EXPECT().IntrospectAccessToken("token").Return(nil, errors.New("unauthorized"))
-
-		_ = ctx.wrapper.VerifyAccessToken(ctx.echoMock, params)
-	})
-
-	t.Run("200 - correct token", func(t *testing.T) {
-		ctx := createContext(t)
-		defer ctx.ctrl.Finish()
-		params := VerifyAccessTokenParams{
-			Authorization: "Bearer token",
-		}
-
-		ctx.echoMock.EXPECT().NoContent(http.StatusOK)
-		ctx.oauthMock.EXPECT().IntrospectAccessToken("token").Return(&services.NutsAccessToken{}, nil)
-
-		_ = ctx.wrapper.VerifyAccessToken(ctx.echoMock, params)
-	})
 }
